+++ parlai/crowdsourcing/tasks/fast_acute/run.py
@@ -30,7 +30,7 @@ from parlai.crowdsourcing.tasks.fast_acute.analysis import (
 from parlai.crowdsourcing.tasks.fast_acute.fast_acute_blueprint import (
     FAST_ACUTE_BLUEPRINT_TYPE,
 )
-from parlai.crowdsourcing.tasks.fast_acute.util import (
+from parlai.crowdsourcing.tasks.acute_eval.util import (
     get_hashed_combo_path,
     ACUTE_EVAL_TASK_DIRECTORY,
     FAST_ACUTE_TASK_DIRECTORY,
