+++ tests/crowdsourcing/tasks/fast_acute/test_no_self_chat_fast_acute.py
@@ -22,7 +22,7 @@ try:
     from parlai.crowdsourcing.tasks.fast_acute.fast_acute_blueprint import (
         FAST_ACUTE_BLUEPRINT_TYPE,
     )
-    from parlai.crowdsourcing.tasks.fast_acute.util import AbstractFastAcuteTest
+    from parlai.crowdsourcing.tasks.acute_eval.util import AbstractFastAcuteTest
 
     class TestNoSelfChatFastAcute(AbstractFastAcuteTest):
         """
